fix decal

+++ common/com_model.h
@@ -143,8 +143,8 @@ struct decal_s
 {
 	decal_t		*pnext;			// linked list for each surface
 	msurface_t	*psurface;		// Surface id for persistence / unlinking
-	short		dx;				// Offsets into surface texture (in texture coordinates, so we don't need floats)
-	short		dy;
+	float		dx;				// Offsets into surface texture (in texture coordinates, so we don't need floats)
+	float		dy;
 	float		scale;		// Pixel scale
 	short		texture;		// Decal texture
 	byte		flags;			// Decal flags
